Removed Next Generation contributions
Apologies to the respective authors but the additional graphics and
dialogues feel too inconsistent with the original content of the Fish
Fillets game. I put the game to its earlier form. All the changes are
bundled in this commit so if anyone wants they're free to revert this in
their own fork, but please do not ask me for reintroducing it in this
distribution.