Small change in death logic, see below
This is to address a special case where:
1. object A falls onto fish X,
2. together with skeleton of X falls onto object B,
3. B is supported by the other fish Y but also other objects.
In the original logic,
- A has SupportType::weak (skeleton)
- thus its fall kills all living obstacles(Direction::down),
- as SupportType::weak stops the traversal in calcSupport to show that
B has SupportType::wall.
Thus the fall of A kills Y via X and B. This rare arrangement happens in
'windoze'.
In the new logic, obstacles() stops at skeletons. Thus,
- A has only obstacle: X,
- X falls on B, getting SupportType::wall, thus avoiding death of Y.
This caused another problem, if Y is just below X as X is killed by A,
Y would be now protected by not being an obstacles() of A. This is fixed
by death() calling evalMotion(Direction::down), as if a killed fish just
fell. If X has nothing below it, evalMotion() has no effect.
One new feature is that X killed by other means, like pushing A wrong
way, also drops onto a fish directly underneath. This was evaluated as
within the spirit of the rules.